Collect code coverage
This adds a code coverage collector for XDebug. It works well except
that framework code is hard to exclude from its own coverage collection.
That means that if we use `expect(subject()->getId())->toBe(42)`, the
XDebug adapter will not only collect coverage for the `getId` method of
the subject but also for the `Expectation` class and the `subject`
variable resolution.

That doesn’t pose a problem for non-framework code though because
`vendor` code and spec files (`.spec.php`) are excluded from coverage
collection through a hard-coded filter.

While I could use PHPUnit’s [`php-code-coverage`][pcc], I don’t see a
point in doing that right now. I’m mostly interested in the code
coverage percentage and not too much in filtering except. See [this
issue][cov-iteration2] for a potential future iteration using the
library due to its in-built features.

I also wanted to use XDebug directly for once. Similar to other parts of
this framework, doing things manually without other dependenies might
not be the best solution but is great for learning. :D

[pcc]: https://github.com/sebastianbergmann/php-code-coverage
[cov-iteration2]: https://gitlab.com/codingpaws/pspec/-/issues/27

If I wanted to support phpdbg or another code coverage collector, I can
simply extend the `src/Coverage/Adapter.php` and register it in the
`Executor::registerCoverageAdapters` method.