Speed up variable resolution by 20%
This speeds up variable resolution by over 20%. It is achieved by
replacing the wasteful top-down variable resolver with an iterative
bottom-up resolver. In total, PSpecs own tests run >12ms faster on my
machine.

Previously, to resolve a variable in the following example, PSpec would:

1. visit all parents, collecting them in a list,
1. reverse the list,
1. walk through all parents reversed, checking for the variable, and
1. if the variable was not found: check the current node’s variables.

```php
describe('add', function () {
  let('a', 5);
  let('b', 0.5);

  context('with a positive integer', function () {
    let('b', 3);

    it('returns a correct value', function () {
      expect(add($this->a, $this->b))->toBe($this->a + $this->b);
    });
  });
});
```

This is wasteful because the parents are visited twice and because the
precedence of a variables increases with the depth in the node tree
(i.e. the more nested a `describe` block is, the more its variables have
priority). To get variable `b` in the test case, we should first look in
the `context` node and not the `describe` one as the top `b` would be
overwritten by the second `b` definition.

By simply iterating over all parents from the leaf to the root, until
we’ve found a variable matching the name, is a lot faster.

This also introduces a secondary improvement. To find a variable on a
node, instead of iterating over all variables, nodes now use a map of
all local variables using an array `name -> Variable`.