Replace bad instance concept with setAsInstance method on PSpec class
Previously, the PSpec class would only allow to be instantiated when
either no PSpec instance was set before or a global value is present.
This is a dirty workaround and has no benefits over this solution, where
the instance is set in the Executor and can be overwritten by calling
`setAsInstance` on a PSpec instance.